move reporting fees to state, it needs to be updated via log emitter handlers
