+++ src/modules/reporting/reducers/reporting-window-stats.js
@@ -1,10 +1,20 @@
 import { UPDATE_REPORTING_WINDOW_STATS } from 'modules/reporting/actions/update-reporting-window-stats'
 import { RESET_STATE } from 'modules/app/actions/reset-state'
+import { formatAttoRep, formatAttoEth } from 'utils/format-number'
 
 const DEFAULT_STATE = {
   startTime: null,
   endTime: null,
   stake: null,
+  reportingFees: {
+    unclaimedEth: formatAttoEth(0, { decimals: 4, zeroStyled: true }),
+    unclaimedRep: formatAttoRep(0, { decimals: 4, zeroStyled: true }),
+    unclaimedForkEth: formatAttoEth(0, { decimals: 4, zeroStyled: true }),
+    unclaimedForkRepStaked: formatAttoRep(0, { decimals: 4, zeroStyled: true }),
+    feeWindows: [],
+    forkedMarket: null,
+    nonforkedMarkets: [],
+  },
 }
 
 export default function (reportingWindowStats = DEFAULT_STATE, action) {
